refactor(strategies): consolidate core retrieval functionality
- Move retrieve_chunks and stream_answer_with_citations from
  src/services/retrieval.py to src/services/strategies/core.py
- Update all strategy imports to use the new consolidated location
- Add comprehensive test coverage for the moved functions
- Maintain identical functionality while improving code organization
